pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = [
    {name = "vitalii-kyiv", email = "v.s.chernous@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] >=0.115.11",
    "sqlalchemy >=2.0.39",
    "alembic >=1.15.1",
    "asyncpg >=0.30.0",
    "pydantic-settings >=2.8.1",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.43.0,<2.0.0)"
]

[tool.poetry.scripts]
dev = "uvicorn src.main:app --reload"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
